requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
asyncpg
pydantic
python-dotenv
psycopg2-binary